Rename file names to lowercase underscore
Follow Dart best practices
