+++ lib/persistence/repositories/app_config_repository.dart
@@ -1,4 +1,4 @@
-import 'package:coffeecard/model/AppConfig.dart';
+import 'package:coffeecard/model/app_config.dart';
 
 abstract class AppConfigRepository {
   Future<AppConfig> getAppConfig();
